+++ app/src/main/java/dev/hossain/trmnl/ui/display/TrmnlMirrorDisplayScreen.kt
@@ -57,7 +57,7 @@ import dev.hossain.trmnl.R
 import dev.hossain.trmnl.data.ImageMetadataStore
 import dev.hossain.trmnl.di.AppScope
 import dev.hossain.trmnl.ui.FullScreenMode
-import dev.hossain.trmnl.ui.config.AppConfigScreen
+import dev.hossain.trmnl.ui.settings.AppConfigScreen
 import dev.hossain.trmnl.ui.refreshlog.DisplayRefreshLogScreen
 import dev.hossain.trmnl.util.CoilRequestUtils
 import dev.hossain.trmnl.util.TokenManager
